+++ frontend/src/metabase/dashboard/containers/AutomaticDashboardApp/AutomaticDashboardApp.jsx
@@ -33,10 +33,12 @@ import { useDashboardUrlQuery } from "../hooks/use-dashboard-url-query";
 import S from "./AutomaticDashboardApp.module.css";
 import { XrayIcon } from "./XrayIcon";
 
-const getDashboardId = (state, { params: { splat }, location: { hash } }) =>
+type AutomaticDashboardAppRouterProps = WithRouterProps<{ split: string }>
+
+const getDashboardId = (_state: State, { params: { splat }, location: { hash } }: AutomaticDashboardAppRouterProps) =>
   `/auto/dashboard/${splat}${hash.replace(/^#?/, "?")}`;
 
-const mapStateToProps = (state, props) => ({
+const mapStateToProps = (state: State, props: AutomaticDashboardAppRouterProps) => ({
   metadata: getMetadata(state),
   dashboardId: getDashboardId(state, props),
   isHeaderVisible: getIsHeaderVisible(state),
@@ -49,7 +51,10 @@ const mapDispatchToProps = {
   addUndo,
 };
 
-class AutomaticDashboardAppInner extends Component {
+const connector = connect(mapStateToProps, mapDispatchToProps)
+type ReduxProps = ConnectedProps<typeof connector>
+
+class AutomaticDashboardAppInner extends Component<ReduxProps & AutomaticDashboardAppRouterProps> {
   state = {
     savedDashboardId: null,
   };
@@ -208,7 +213,7 @@ class AutomaticDashboardAppInner extends Component {
 }
 
 export const AutomaticDashboardAppConnected = _.compose(
-  connect(mapStateToProps, mapDispatchToProps),
+  connector,
   DashboardData,
   title(({ dashboard }) => dashboard && dashboard.name),
 )(AutomaticDashboardAppInner);
@@ -220,88 +225,6 @@ const TransientTitle = ({ dashboard }) =>
     <span>{dashboard.name}</span>
   ) : null;
 
-const RELATED_CONTENT = {
-  compare: {
-    get title() {
-      return t`Compare`;
-    },
-    icon: "compare",
-  },
-  "zoom-in": {
-    get title() {
-      return t`Zoom in`;
-    },
-    icon: "zoom_in",
-  },
-  "zoom-out": {
-    get title() {
-      return t`Zoom out`;
-    },
-    icon: "zoom_out",
-  },
-  related: {
-    get title() {
-      return t`Related`;
-    },
-    icon: "connections",
-  },
-};
-
-const SuggestionsList = ({ suggestions, section }) => (
-  <Box component="ol" my="sm">
-    {Object.keys(suggestions).map((s, i) => (
-      <li key={i} className={CS.my2}>
-        <SuggestionSectionHeading>
-          {RELATED_CONTENT[s].title}
-        </SuggestionSectionHeading>
-        {suggestions[s].length > 0 &&
-          suggestions[s].map((item, itemIndex) => (
-            <Link
-              key={itemIndex}
-              to={item.url}
-              className={cx(CS.hoverParent, CS.hoverVisibility, S.ItemLink)}
-            >
-              <Card className={CS.p2} hoverable>
-                <Flex align="center">
-                  <Icon
-                    name={RELATED_CONTENT[s].icon}
-                    color={color("accent4")}
-                    className={CS.mr1}
-                  />
-                  <h4 className={CS.textWrap}>{item.title}</h4>
-                  <Box ml="auto" className={CS.hoverChild}>
-                    <Tooltip label={item.description}>
-                      <Icon name="info_outline" color={color("bg-dark")} />
-                    </Tooltip>
-                  </Box>
-                </Flex>
-              </Card>
-            </Link>
-          ))}
-      </li>
-    ))}
-  </Box>
-);
-
-const SuggestionSectionHeading = ({ children }) => (
-  <h5
-    style={{
-      fontWeight: 900,
-      textTransform: "uppercase",
-      color: color("text-medium"),
-    }}
-    className={CS.mb1}
-  >
-    {children}
-  </h5>
-);
-
-const SuggestionsSidebar = ({ related }) => (
-  <Flex direction="column" py="md" px="xl">
-    <Title py="sm" px={0} order={2}>{t`More X-rays`}</Title>
-    <SuggestionsList suggestions={related} />
-  </Flex>
-);
 
 // Workaround until AutomaticDashboardApp is refactored to be a function component
 // (or even better, merged/generalized with DashboardApp)
