+++ frontend/src/metabase/dashboard/containers/AutomaticDashboardApp/index.ts
@@ -1,0 +1,1 @@
+export * from "./AutomaticDashboardApp"
